Retrieve current git commit hash #83
rename get_branch_name to git_info module
